requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
asgiref=3.4.1=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
beautifulsoup4=4.10.0=pypi_0
ca-certificates=2021.10.26=haa95532_2
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py39haa95532_0
charset-normalizer=2.0.9=pypi_0
crispy-bootstrap5=0.6=pypi_0
cycler=0.11.0=pypi_0
dj-database-url=0.5.0=pypi_0
django=3.2.5=pyhd3eb1b0_0
django-bootstrap-v5=1.0.7=pypi_0
django-cleanup=5.2.0=pypi_0
django-crispy-forms=1.13.0=pypi_0
django-heroku=0.3.1=pypi_0
flatbuffers=2.0=pypi_0
fonttools=4.28.4=pypi_0
gast=0.4.0=pypi_0
google-auth=2.3.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.43.0=pypi_0
h5py=3.6.0=pypi_0
heroku=0.1.4=pypi_0
idna=3.3=pypi_0
imageio=2.13.3=pypi_0
importlib-metadata=4.9.0=pypi_0
keras=2.7.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.2=pypi_0
krb5=1.19.2=h5b6d351_0
libclang=12.0.0=pypi_0
libpq=12.9=hb652d5d_1
markdown=3.3.6=pypi_0
matplotlib=3.5.1=pypi_0
networkx=2.6.3=pypi_0
numpy=1.21.4=pypi_0
oauthlib=3.1.1=pypi_0
opencv-contrib-python=4.5.4.60=pypi_0
opencv-python=4.5.4.60=pypi_0
openssl=1.1.1l=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pypi_0
pillow=8.4.0=pypi_0
pip=21.2.4=py39haa95532_0
protobuf=3.19.1=pypi_0
psycopg2=2.8.6=py39hcd4344a_1
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.0.6=pypi_0
python=3.9.7=h6244533_1
python-dateutil=2.8.2=pypi_0
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.2.0=pypi_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.8=pypi_0
scikit-image=0.19.1=pypi_0
scipy=1.7.3=pypi_0
setuptools=58.0.4=py39haa95532_0
six=1.16.0=pypi_0
soupsieve=2.3.1=pypi_0
sqlite=3.36.0=h2bbff1b_0
sqlparse=0.4.1=py_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.7.0=pypi_0
tensorflow-estimator=2.7.0=pypi_0
tensorflow-io-gcs-filesystem=0.23.1=pypi_0
termcolor=1.1.0=pypi_0
tifffile=2021.11.2=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
urllib3=1.26.7=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
whitenoise=5.3.0=pypi_0
wincertstore=0.2=py39haa95532_2
wrapt=1.13.3=pypi_0
zipp=3.6.0=pypi_0
zlib=1.2.11=h8cc25b3_4
